outsourced report creation to a new class
