+++ src/Status.cxx
@@ -1,5 +1,3 @@
-#include <map>
-
 #include "Status.hpp"
 
 #include <rapidjson/document.h>
@@ -48,6 +46,15 @@ void FileWriter::Status::StreamerStatusType::reset() {
 
 ////////////////////////////////
 // StreamerStatus implementation
+
+double average(const double& sum, const double& N) {
+  return sum/N;
+}
+
+double standard_deviation(const double& sum, const double& sum_squared, const double& N) {
+  return std::sqrt( (N*sum_squared - sum*sum)/(N*(N-1)) );
+}
+
 FileWriter::Status::StreamerStatusType
 FileWriter::Status::StreamerStatus::fetch_status() {
   std::unique_lock<std::mutex> lock(m);
@@ -58,8 +65,8 @@ void FileWriter::Status::StreamerStatus::add_message(const double &bytes) {
   std::unique_lock<std::mutex> lock(m);
   current.bytes += bytes;
   current.bytes_squared += bytes * bytes;
-  current.messages += 1;
-  current.messages_squared += 1;
+  current.messages++;
+  current.messages_squared++;
 }
 
 FileWriter::Status::StreamerStatisticsType
@@ -69,24 +76,18 @@ FileWriter::Status::StreamerStatus::fetch_statistics() {
   using namespace std::chrono;
 
   system_clock::time_point t = system_clock::now();
-  auto count = duration_cast<seconds>(t - last_time).count();
   if (current.messages > 0) {
-    st.average_message_size = current.bytes / current.messages;
-    st.standard_deviation_message_size =
-        std::sqrt((current.bytes_squared / current.messages -
-                   st.average_message_size * st.average_message_size) /
-                  current.messages);
+    st.average_message_size = average(current.bytes,current.messages);
+    st.standard_deviation_message_size = standard_deviation(current.bytes, current.bytes_squared, current.messages);
 
-    st.average_message_frequency = current.messages / count;
-    st.standard_deviation_message_frequency = std::sqrt(
-        (current.messages / count -
-         st.average_message_frequency * st.average_message_frequency) /
-        count);
+    double elapsed_time = duration_cast<seconds>(t - last_time).count();
+    st.average_message_frequency = average(current.messages, elapsed_time);
+    st.standard_deviation_message_frequency = standard_deviation(current.messages, current.messages_squared, elapsed_time);
   } else {
-    st.average_message_size = -1.;
-    st.standard_deviation_message_size = -1.;
-    st.average_message_frequency = -1.;
-    st.standard_deviation_message_frequency = -1.;
+    st.average_message_size = 0.;
+    st.standard_deviation_message_size = 0.;
+    st.average_message_frequency = 0.;
+    st.standard_deviation_message_frequency = 0.;
   }
   last += current;
   last_time = t;
@@ -94,47 +95,15 @@ FileWriter::Status::StreamerStatus::fetch_statistics() {
   return st;
 }
 
-static const std::map<const int, std::string> stream_master_error_lookup_{
-    {FileWriter::Status::StreamMasterErrorCode::not_started, "not_started"},
-    {FileWriter::Status::StreamMasterErrorCode::running, "running"},
-    {FileWriter::Status::StreamMasterErrorCode::has_finished, "has_finished"},
-    {FileWriter::Status::StreamMasterErrorCode::statistics_failure,
-     "statistics_failure"},
-    {FileWriter::Status::StreamMasterErrorCode::streammaster_error,
-     "streammaster_error"}};
+//////////////////////
+// StreamMasterStatus
 
-static const std::map<const int, std::string> streamer_error_lookup_{
-    {FileWriter::Status::StreamerErrorCode::writing, "writing"},
-    {FileWriter::Status::StreamerErrorCode::stopped, "stopped"},
-    {FileWriter::Status::StreamerErrorCode::configuration_error,
-     " configuration_error"},
-    {FileWriter::Status::StreamerErrorCode::consumer_error, " consumer_error"},
-    {FileWriter::Status::StreamerErrorCode::metadata_error, "metadata_error"},
-    {FileWriter::Status::StreamerErrorCode::topic_partition_error,
-     "topic_partition_error"},
-    {FileWriter::Status::StreamerErrorCode::assign_error, "assign_error"},
-    {FileWriter::Status::StreamerErrorCode::topic_error, "topic_error"},
-    {FileWriter::Status::StreamerErrorCode::offset_error, "offset_error"},
-    {FileWriter::Status::StreamerErrorCode::start_time_error,
-     "start_time_error"},
-    {FileWriter::Status::StreamerErrorCode::message_error, "message_error"},
-    {FileWriter::Status::StreamerErrorCode::write_error, "write_error"},
-    {FileWriter::Status::StreamerErrorCode::not_initialized,
-     "not_initialized"}};
-
-const std::string
-FileWriter::Status::Err2Str(const FileWriter::StreamMasterError &error) {
-  auto it = stream_master_error_lookup_.find(error.value());
-  if (it != stream_master_error_lookup_.end()) {
-    return it->second;
-  }
-  return "Unknown error code";
-};
-const std::string
-FileWriter::Status::Err2Str(const FileWriter::StreamerError &error) {
-  auto it = stream_master_error_lookup_.find(error.value());
-  if (it != stream_master_error_lookup_.end()) {
-    return it->second;
-  }
-  return "Unknown error code";
-};
+FileWriter::Status::StreamMasterStatus& 
+FileWriter::Status::StreamMasterStatus::push(const std::string &topic_name,
+					     const FileWriter::Status::StreamerStatusType &status,
+					     const FileWriter::Status::StreamerStatisticsType &stats) {
+  topic.push_back(topic_name);
+  streamer_status.push_back(status);
+  streamer_stats.push_back(stats);
+  return *this;
+}
